List mode non-pattern icons
Non-pattern iotas in list mode will now be displayed as various symbols according to their type. [] for lists, ⟨⟩ for vectors, # for numbers, ? for widgets, and @ for anything else, which is assumed to be the name of a player or entity. Also, the displayed list of patterns will now be wrapped in intro/retro.